fixed float64 bug.

+++ dataProcess/browsers.go
@@ -49,7 +49,7 @@ func BrowswersCount(c *gin.Context, db *sql.DB) {
 	infos, total := getBrowsersInfo(db, s)
 
 	for i, v := range infos {
-		infos[i].Rate = strconv.FormatFloat(float64(v.Num/total), 'f', 2, 64)
+		infos[i].Rate = strconv.FormatFloat(float64(v.Num)/float64(total), 'f', 2, 64)
 	}
 
 	cd.Rows = infos
